Fix EFI boot on certain non-Chromebook platforms.
Some non-chromebook platforms cannot EFI boot Chromium OS due to "cannot
allocate protected mode pages" error. This is a very old grub2 bug, and
is worked around in this change. A better fix is to upgrade to a newer
version of grub2, but this has caused big problems in the past. This
change is relatively risk-free because it only takes effect if EFI boot
were to fail.

BUG=32776.
TEST=build image with new grub-mkimage, EFI boot on platform.

Change-Id: Ic332f3af3a771d084189dca7fc2f88b281e653b3
Reviewed-on: https://gerrit.chromium.org/gerrit/28043
Reviewed-by: Micah Catlin <micahc@chromium.org>
Commit-Ready: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-by: Shawn Nematbakhsh <shawnn@chromium.org>
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>